+++ Assets/Scripts/Player/SliderBar.cs
@@ -9,7 +9,7 @@ public class SliderBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public virtual void SetMaxValue(int value)
+    public virtual void SetMaxValue(float value)
     {
 
         slider.maxValue = value;
@@ -17,7 +17,7 @@ public class SliderBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public virtual void SetValue(int value)
+    public virtual void SetValue(float value)
     {
         slider.value = value;
         fill.color = gradient.Evaluate(slider.normalizedValue);
