+++ Assets/Scripts/Player/SprintBar.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class SprintBar : SliderBar
 {
 
-    public override void SetMaxValue(int value)
+    public override void SetMaxValue(float value)
     {
         base.SetMaxValue(value);
     }
 
-    public override void SetValue(int value)
+    public override void SetValue(float value)
     {
         base.SetValue(value);
     }
